Fix update_hash_state and setup_hash_state to use readonly fields.
The C sources indicate that the first field of the `s2n_hash` struct
is a const pointer; this change makes the saw specs match.

The old definition of `update_hash_state` was actually incorrect, in
that it specified the `impl` field of the `s2n_hash` struct to be freshly
allocated every time the struct was updated. But in actuality, the
pointer value was usually copied from some other input.

A new pull request (GaloisInc/saw-script#826) makes saw-script more
strict about checking that fresh pointers are actually fresh, which
revealed the bug in the HMAC proofs. This change makes the proofs pass
the more stringent check.